Remove unused, rename and move packages
