+++ app/src/main/java/xyz/gozdzik/cryptolist/MainActivity.kt
@@ -5,9 +5,9 @@ import android.os.Bundle
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class DemoActivity : AppCompatActivity() {
+class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_demo)
+        setContentView(R.layout.activity_main)
     }
 }
