+++ app/src/main/java/xyz/gozdzik/cryptolist/data/local/repository/CurrencyInfoLocalRepository.kt
@@ -2,7 +2,7 @@ package xyz.gozdzik.cryptolist.data.local.repository
 
 import xyz.gozdzik.cryptolist.core.DataDomainMapper
 import xyz.gozdzik.cryptolist.data.local.database.CurrencyInfoDao
-import xyz.gozdzik.cryptolist.data.model.CurrencyInfoDto
+import xyz.gozdzik.cryptolist.data.local.model.CurrencyInfoDto
 import xyz.gozdzik.cryptolist.domain.model.CurrencyInfo
 import xyz.gozdzik.cryptolist.domain.repository.CurrencyInfoRepository
 import javax.inject.Inject
