+++ app/src/main/res/navigation/nav_graph.xml
@@ -3,16 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/demoFragment">
-    <fragment
-        android:id="@+id/demoFragment"
-        android:name="xyz.gozdzik.cryptolist.presentation.demo.DemoFragment"
-        android:label="DemoFragment"
-        tools:layout="@layout/fragment_demo" >
-        <action
-            android:id="@+id/navigateToFragmentCurrencyList"
-            app:destination="@id/fragmentCurrencyList" />
-    </fragment>
+    app:startDestination="@id/fragmentCurrencyList">
     <fragment
         android:id="@+id/fragmentCurrencyList"
         android:name="xyz.gozdzik.cryptolist.presentation.currencylist.CurrencyListFragment"
